fix routes router file
